requirements.txt
huggingface-hub==0.25.0
numpy==1.24.4
pickle
pandas==2.0.3
Pillow==10.0.0
pyiqa==0.1.12
pytorch-lightning==1.4.2
scipy==1.10.1
thop
torch==2.2.1
torchvision==0.17.1
transformers==4.37.2
typed-argument-parser==1.10.1
wandb==0.17.8
xformers==0.0.24
pytz
